refactor:  move error handling  and related constants to common errors module
