+++ src/jobs/models/errors.ts
@@ -4,17 +4,3 @@ export class JobNotFoundError extends Error {
     Object.setPrototypeOf(this, JobNotFoundError.prototype);
   }
 }
-
-export class InvalidUpdateError extends Error {
-  public constructor(message: string) {
-    super(message);
-    Object.setPrototypeOf(this, InvalidUpdateError.prototype);
-  }
-}
-
-export const prismaKnownErrors = {
-  /**An operation failed because it depends on one or more records that were required but not found. {cause} */
-  recordNotFound: 'P2025',
-} as const;
-
-export const BAD_STATUS_CHANGE = 'Invalid status change';
